Fixes #1, added parsing capability of xyz file. Also enriched the list of elements slightly

+++ SymmLab/SLModelLine.h
@@ -10,6 +10,6 @@
 
 @interface SLModelLine : SLAbstractModel
 
-- (id) initWithPoints:(GLKVector3 *)points colors:(SLColor *)colors count:(NSUInteger)count;
+- (id) initWithPoints:(GLKVector3 *)points colors:(SLColor *)colors count:(GLuint)count;
 
 @end
